fix docs and addressing more comments
